+++ atn-atm-mediator/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added Sha256 hash of DID to Session Struct. Optimisation
 * Changed Mediator GLOBAL Stats from u64 to i64 in case of negative values
 * Removed unused common/Stats module
+* improved delete_message return type to DeleteMessageResponse Struct, with returns for success and errors
 
 ### 6th June 2024