+++ atn-atm-sdk/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Changelog history
 
+### 7th June 2024
+
+* improved delete_message return type to DeleteMessageResponse Struct, with returns for success and errors
+
 ### 6th June 2024
 
 * Modified list_messages() to send the requested did as part of the request.